restructure and cleaning
