+++ CORDEX-CMIP6_fixed.json
@@ -1,0 +1,22 @@
+{
+    "fixed": {
+        "mip_era": [
+            "CMIP6"
+        ],
+        "product": [
+            "model-output"
+        ],
+        "license": [
+            "https://cordex.org/data-access/cordex-cmip6-data/cordex-cmip6-terms-of-use"
+        ],
+        "tracking_id": [
+            "hdl:21.14103/.*"
+        ],
+        "variant_label": [
+            "r[[:digit:]]\\{1,\\}i[[:digit:]]\\{1,\\}p[[:digit:]]\\{1,\\}f[[:digit:]]\\{1,\\}$"
+        ],
+        "Conventions": [
+            "CF-1.11"
+        ]
+    }
+}
