+++ scripts/cordex_cv/cli.py
@@ -1,6 +1,16 @@
-from .tools import create_cv as cv
+from . import create_cv, create_domain_id, create_driving_source_id
 
 
-def create_cv():
+def cv():
     print("creating CV")
-    cv()
+    create_cv()
+
+
+def domain_id():
+    print("creating domain_id")
+    create_domain_id()
+
+
+def driving_source_id():
+    print("creating driving_source_id")
+    create_driving_source_id()
