+++ scripts/pyproject.toml
@@ -11,6 +11,8 @@ version = "0.1.0"
 
 [project.scripts]
 create-cv = "cordex_cv.cli:create_cv"
+create-domain_id = "cordex_cv.cli:domain_id"
+create-driving_source_id = "cordex_cv.cli:driving_source_id"
 
 [tool.setuptools]
 packages = ["cordex_cv"]
