+++ scripts/tests.py
@@ -1,9 +1,6 @@
 import json
 import pytest
-import sys
-import os
 
-sys.path.insert(0, os.path.join(os.getcwd(), "tools"))
 from scripts.tools import update_table
